requirements.txt
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.123.8
h11==0.16.0
idna==3.11
Jinja2==3.1.4
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.29
starlette==0.50.0
typing_extensions==4.15.0
uvicorn==0.38.0
